feat(crypto-c99): §3.18 ML-DSA increment 2 — rounding/hint layer + self-audit fixes
PQ track increment 2 on top of R63's NTT: the FIPS 204 coefficient rounding that
keygen and sign/verify sit on. src/crypto/mldsa/rounding.{c,h}:
- power2round  — t = t1*2^D + t0 (the keygen public-key split)
- decompose    — HighBits/LowBits around the GAMMA2 grid (the signing round)
- make_hint / use_hint — the signature's per-coefficient carry hint
gamma2 is a runtime arg (GAMMA2_88 for ML-DSA-44, GAMMA2_32 for ML-DSA-65/87),
so one core serves all three sets. Every function proven in python over 200k
random inputs/gamma2 BEFORE the C was written. test-mldsa-c99 gains 5 rounding
assertions: power2round/decompose reconstruction + bounds, the use_hint semantic
round-trip use_hint(r,[HB(r)!=HB(r+z)])==HB(r+z), make_hint's definitional
formula, and boundary KATs at the decomposition seams (both gamma2). Additive;
still no signer.

Also remediates the 4 CONFIRMED findings from the adversarial self-audit of R63
(4-lens Workflow, verify-to-refute):
- [medium/test-strength] the forward-NTT byte-exact KAT (ntt_out) had NO oracle
  independent of the transform under test — the round-trip AND the schoolbook
  convolution are both invariant under a consistent permutation of the NTT domain,
  so a symmetric zetas-ordering bug passed the entire gate. FIX: an independent
  direct-DFT oracle ntt(X)[j] == root^(2*brv8(j)+1) (closed-form root evaluation,
  no butterfly/zetas reuse) added to the C self-test AND both file-side gates
  (verify_mldsa_vectors.py + test_c99_vector_files.sh); the false "second oracle
  implicitly covered" comment removed.
- [low/UB, x2] reduce32 did (a + (1<<22)) in int32 — signed-overflow UB for a
  near INT32_MAX, contradicting the "any int32 a" header contract. FIX: widen the
  bias-add to int64 (UB-free for any int32, costs nothing).

Build: determ + determ-wallet + determ-light all green (MSVC). test-mldsa-c99
12 assertions PASS (incl. the new independent root-ordering KAT). FAST 175/175.
LESSON: round-trip + convolution self-tests can BOTH be blind to a self-consistent
transform bug; a byte-exact interop KAT needs an oracle whose expected values do
not come from the implementation under test.

+++ src/crypto/mldsa/reduce.c
@@ -20,10 +20,14 @@ int32_t determ_mldsa_montgomery_reduce(int64_t a) {
     return t;
 }
 
-/* Barrett: t = round(a / q) via the fixed-point (a + 2^22) >> 23, then a - t*q. */
+/* Barrett: t = round(a / q) via the fixed-point (a + 2^22) >> 23, then a - t*q.
+ * The bias add is done in int64 so the contract holds for ANY int32 a without
+ * signed-overflow UB (the canonical reference assumes a <= 2^31 - 2^22; widening
+ * removes that precondition at no cost — t and a - t*q both stay in int32 range
+ * because t = round(a/q) is bounded by |a|/q + 1). */
 int32_t determ_mldsa_reduce32(int32_t a) {
     int32_t t;
-    t = (a + (1 << 22)) >> 23;
+    t = (int32_t)(((int64_t)a + (1 << 22)) >> 23);
     t = a - t * Q;
     return t;
 }
